@@ -1,18 +1,34 @@
-export const FETCH_POSTS_REQUEST = "FETCH_POSTS_REQUEST";
-export const FETCH_POSTS_SUCCESS = "FETCH_POSTS_SUCCESS";
-export const FETCH_POSTS_FAILURE = "FETCH_POSTS_FAILURE";
+// src/store/actions.js
+export const FETCH_START = "FETCH_START";
+export const FETCH_SUCCESS = "FETCH_SUCCESS";
+export const FETCH_ERROR = "FETCH_ERROR";
 
-export const fetchPosts = () => {
-  return async (dispatch) => {
-    dispatch({ type: FETCH_POSTS_REQUEST });
+export const fetchLoremStart = () => ({
+  type: FETCH_START,
+});
 
-    try {
-      const res = await fetch("https://api.lorem.com/ipsum");
-      const data = await res.json();
+export const fetchLoremSuccess = (data) => ({
+  type: FETCH_SUCCESS,
+  payload: data,
+});
 
-      dispatch({ type: FETCH_POSTS_SUCCESS, payload: data });
-    } catch (err) {
-      dispatch({ type: FETCH_POSTS_FAILURE, payload: err.message });
-    }
-  };
+export const fetchLoremError = (error) => ({
+  type: FETCH_ERROR,
+  payload: error,
+});
+
+// Mock thunk — no real API call
+export const fetchLorem = () => async (dispatch) => {
+  dispatch(fetchLoremStart());
+
+  // small delay to show "Loading..."
+  await new Promise((res) => setTimeout(res, 600));
+
+  // fake payload that matches the required shape
+  dispatch(
+    fetchLoremSuccess({
+      title: "Lorem Ipsum",
+      body: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed non risus.",
+    })
+  );
 };
