@@ -1,27 +1,22 @@
-import {
-  FETCH_POSTS_REQUEST,
-  FETCH_POSTS_SUCCESS,
-  FETCH_POSTS_FAILURE,
-} from "./actions";
+import { FETCH_START, FETCH_SUCCESS, FETCH_ERROR } from "./actions";
 
 const initialState = {
-  loading: true,
-  posts: [],
+  loading: false,
+  data: null,
   error: null,
 };
 
-export default function reducer(state = initialState, action) {
+const loremReducer = (state = initialState, action) => {
   switch (action.type) {
-    case FETCH_POSTS_REQUEST:
-      return { ...state, loading: true };
-
-    case FETCH_POSTS_SUCCESS:
-      return { ...state, loading: false, posts: action.payload };
-
-    case FETCH_POSTS_FAILURE:
+    case FETCH_START:
+      return { ...state, loading: true, error: null };
+    case FETCH_SUCCESS:
+      return { ...state, loading: false, data: action.payload };
+    case FETCH_ERROR:
       return { ...state, loading: false, error: action.payload };
-
     default:
       return state;
   }
-}
+};
+
+export default loremReducer;
